reuse URLUtilities
