+++ RevenueCatUI/CustomerCenter/URLUtilities.swift
@@ -16,9 +16,10 @@
 import Foundation
 import SwiftUI
 
-enum URLUtilities {
-
 #if os(iOS)
+import UIKit
+
+enum URLUtilities {
 
     @available(iOS 15.0, macOS 12.0, tvOS 15.0, watchOS 8.0, *)
     @available(macOS, unavailable)
@@ -43,13 +44,46 @@ enum URLUtilities {
 
         components.queryItems = queryItems
 
-        guard let url = components.url, URLOpener.canOpenURL(url) else {
+        guard let url = components.url, Self.canOpenURL(url) else {
             return nil
         }
 
         return url
     }
 
-#endif
+    static func openURLIfNotAppExtension(_ url: URL) {
+        guard !Self.isAppExtension,
+              let application = Self.sharedUIApplication else {
+            return
+        }
+
+        let selector = NSSelectorFromString("openURL:options:completionHandler:")
+        typealias ClosureType = @convention(c) (AnyObject, Selector, NSURL, NSDictionary?, Any?) -> Void
+        let methodIMP: IMP! = application.method(for: selector)
+        let openURLMethod = unsafeBitCast(methodIMP, to: ClosureType.self)
+        openURLMethod(application, selector, url as NSURL, nil, nil)
+    }
+
+    static func canOpenURL(_ url: URL) -> Bool {
+        guard !Self.isAppExtension,
+              let application = Self.sharedUIApplication else {
+            return false
+        }
+        return application.canOpenURL(url)
+    }
 
 }
+
+private extension URLUtilities {
+
+    static var isAppExtension: Bool {
+        Bundle.main.bundlePath.hasSuffix(".appex")
+    }
+
+    static var sharedUIApplication: UIApplication? {
+        return UIApplication.value(forKey: "sharedApplication") as? UIApplication
+    }
+
+}
+
+#endif
